Added gas threshold functions for lenticular plots
The first function has a hard threshold, below which nothing becomes a
lenticular galaxy. The second function has a soft threshold, so it
becomes increasingly less likely that it becomes a lenticular the
further below the threshold it is.

I made them 2 seperate functions just to make it easier to read and
comment out if necessary.