Added function to make lenticulars without mergers
The function uses the growth of many galaxies and labels a galaxy as a
lenticular if it has grown to a certain fraction of its final mass
within a certain redshift time. However, this is currently not working
as it should because reshift doesn't scale linearly to lookback time. I
need to fix this by using the Colossus module to convert redshift to
lookback time and use that to define the cutoff parameters.