Fixed Sersic index plot
Previously, I was trying to caluclate Sersic Index from other quantities
present in the dataframe. But I noticed that there is a n_bulge header
which is the Sersic Index of the bulge. I am now using this to calculate
Sersic Index for each morphology at different stellar masses. This gives
Sersic Indices that are more probable.