Added several models including Bulge growth model
Added several models including a simplified Cook Model, some models I
made up and was testing and the final model of my internship: the bulge
growth model.

This  bulge growth model assigns a bulge mass to each galaxy and if the
BT ratio is greater than some threshold, that galaxy is included as a
lenticular galaxy. This is the final model I worked on during my summer
placement.